fix(Concat): Enables nested-concat as second arg
Prior, we assumed that there must be exactly two functions on the functionBuilder stack when parsing concat.
However, `CONCAT(fieldOne, CONCAT(VARSIZED(" "), fieldTwo))` is perfectly valid, but leads to a situation where there are three functions on the stack (1. FieldAccess(fieldOne), 2. Constant(VARSIZED(" ")), 3. FieldAccess(fieldTwo)). Thus, we removed the check.